modify the post route with JWT validation

+++ src/controllers/PostController.js
@@ -21,7 +21,7 @@ const getAllPosts = async (request, response) => {
 
 const getSpecificUserPosts = async (request, response) => {
   try {
-    const allPosts = await Post.find({ userId: request.params.userId }).exec();
+    const allPosts = await Post.find({ userId: request.headers.userId }).exec();
     response.json({
       data: allPosts
     });
@@ -54,7 +54,6 @@ const createPost = async (request, response) => {
   try {
     const file = request.file;
     const photos = await uploadFileToS3(file);
-    console.log(request.body);
     const newPost = new Post({
       name: request.body.name,
       species: request.body.species,
@@ -65,7 +64,7 @@ const createPost = async (request, response) => {
       suburb: request.body.suburb,
       contactInfo: request.body.contactInfo,
       status: request.body.status,
-      userId: request.params.userId
+      userId: request.headers.userId
     });
     const savedPost = await newPost.save();
     response.json({
@@ -79,33 +78,20 @@ const createPost = async (request, response) => {
 };
 
 const deletePost = async (request, response) => {
-  // try {
-  //   const post = await Post.findById(request.params.postId).exec();
-
-  //   // only user who created the post or admin can delete the post
-  //   // user equals to find if the user is the same as the one who created the post
-  //   // can not use === because you can not compare two objects, they are always different
-  //   if (post.userId.equals(request.headers.userId) || request.headers.role == "admin") {
-  //     const deletedPost = await Post.findByIdAndDelete(request.params.postId).exec();
-  //     response.json({
-  //       data: deletedPost
-  //     });
-  //   } else {
-  //     throw new Error("You are not authorized to delete this post.");
-  //   }
-  // } catch (error) {
-  //   response.json({
-  //     error: error.message
-  //   });
-  // }
   try {
-    const deletedPost = await Post.findByIdAndDelete(request.params.postId).exec();
-    if (!deletedPost) {
-      throw new Error("Post not found");
+    const post = await Post.findById(request.params.postId).exec();
+
+    // only user who created the post or admin can delete the post
+    // user equals to find if the user is the same as the one who created the post
+    // can not use === because you can not compare two objects, they are always different
+    if (post.userId.equals(request.headers.userId) || request.headers.role == "admin") {
+      const deletedPost = await Post.findByIdAndDelete(request.params.postId).exec();
+      response.json({
+        data: deletedPost
+      });
+    } else {
+      throw new Error("You are not authorized to delete this post.");
     }
-    response.json({
-      data: deletedPost
-    });
   } catch (error) {
     response.json({
       error: error.message
@@ -114,47 +100,32 @@ const deletePost = async (request, response) => {
 };
 
 const updatePost = async (request, response) => {
-  // try {
-  //   const post = await Post.findById(request.params.postId).exec();
-  //   if (!post.userId.equals(request.headers.userId)) {
-  //     throw new Error("You are not authorized to update this post.");
-  //   } else {
-  //     const updatedPost = await Post.findByIdAndUpdate(
-  //       request.params.postId,
-  //       request.body.updatedData,
-  //       { new: true }
-  //     ).exec();
-  //     response.json({
-  //       data: updatedPost
-  //     });
-  //   }
-  // } catch (error) {
-  //   response.json({
-  //     error: error.message
-  //   });
-  // }
   try {
     const post = await Post.findById(request.params.postId).exec();
     if (!post) {
       throw new Error("Post not found");
     }
-    const updatedData = {
-      name: request.body.name || post.name,
-      species: request.body.species || post.species,
-      breed: request.body.breed || post.breed,
-      color: request.body.color || post.color,
-      description: request.body.description || post.description,
-      suburb: request.body.suburb || post.suburb,
-      contactInfo: request.body.contactInfo || post.contactInfo,
-      status: post.status
-    };
+    if (!post.userId.equals(request.headers.userId)) {
+      throw new Error("You are not authorized to update this post.");
+    } else {
+      const updatedData = {
+        name: request.body.name || post.name,
+        species: request.body.species || post.species,
+        breed: request.body.breed || post.breed,
+        color: request.body.color || post.color,
+        description: request.body.description || post.description,
+        suburb: request.body.suburb || post.suburb,
+        contactInfo: request.body.contactInfo || post.contactInfo,
+        status: post.status
+      };
 
-    const updatedPost = await Post.findByIdAndUpdate(request.params.postId, updatedData, {
-      new: true
-    }).exec();
-    response.json({
-      data: updatedPost
-    });
+      const updatedPost = await Post.findByIdAndUpdate(request.params.postId, updatedData, {
+        new: true
+      }).exec();
+      response.json({
+        data: updatedPost
+      });
+    }
   } catch (error) {
     response.json({
       error: error.message
